backend/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
cmdstanpy==1.2.0
contourpy==1.1.1
cycler==0.12.1
Django==4.2.5
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fonttools==4.43.1
frozendict==2.3.8
gunicorn==21.2.0
holidays==0.35
html5lib==1.1
idna==3.4
importlib-resources==6.1.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
multitasking==0.0.11
numpy==1.26.0
packaging==23.2
pandas==2.1.1
peewee==3.16.3
Pillow==10.1.0
prophet==1.1.5
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stanio==0.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
webencodings==0.5.1
yfinance==0.2.30
